requirements.txt
Django==1.9.6
django-annoying==0.9.0
django-bower==5.1.0
django-datetime-widget==0.9.3
djangorestframework==3.3.3
future==0.15.2
gunicorn==19.6.0
pefile==2016.3.28
PyInstaller==3.2
pypiwin32==219
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
virtualenv==15.0.2
whitenoise==3.2
